wrong use of @stimulus.schema
Creation of  Tones_Grating table is wrong beacuse it is actualy using tables from Tones and Gratings

Reason i check it is beacuse it creates an error on the creation of the tables because of the Tones_Grating is not CamelCase. It is Also creates empty tables with no use.